moving settlors and protectors specs and mapper objects to their right location within the solution
