+++ app/uk/gov/hmrc/trustapi/mapping/todes/DesProtectorsMapper.scala
@@ -14,15 +14,15 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.common.mapping.todes
+package uk.gov.hmrc.trustapi.mapping.todes
 
-import uk.gov.hmrc.common.des.DesSettlor
-import uk.gov.hmrc.common.rest.resources.core.Individual
+import uk.gov.hmrc.common.des.DesProtectorType
+import uk.gov.hmrc.trustapi.rest.resources.core.Protectors
 
 
-object DesSettlorMapper {
-  def toDes(individual: Individual) : DesSettlor = {
-    DesSettlor(DesNameMapper.toDes(individual),individual.dateOfBirth,DesIdentificationMapper.toDes(individual))
+object DesProtectorsMapper {
+  def toDes(protector: Protectors) : DesProtectorType = {
+    DesProtectorType(protector.individuals.map(li=>li.map(i=>DesProtectorMapper.toDes(i))),
+      protector.companies.map(lc => lc.map(c=>DesProtectorCompanyMapper.toDes(c))))
   }
 }
-
